+++ app/src/main/java/com/nbcamp/tripgo/data/model/festivals/Body.kt
@@ -1,4 +1,4 @@
-package com.nbcamp.tripgo.data.model.festival
+package com.nbcamp.tripgo.data.model.festivals
 
 
 import com.google.gson.annotations.SerializedName
